Add jwt authorization for API

+++ controllers/api/v1/post.js
@@ -20,18 +20,28 @@ module.exports.post = async function(req, res){
     }
     catch(err){
         console.log(err);
-        return;
+        return res.status(404).json({
+            message : "Not found"
+        });
     } 
 }
 
 module.exports.destroy =async function(req, res){
     try{
         let post = await Post.findById(req.params.id);
-        post.remove();
-        Comment.deleteMany({post : req.params.id});
-        return res.status(200).json({
-            message : 'post and associated comments deleted'
-        });
+        if(post.user == req.user.id){
+            post.remove();
+            Comment.deleteMany({post : req.params.id});
+            return res.status(200).json({
+                message : 'post and associated comments deleted'
+            });
+        }
+        else{
+            return res.status(401).json({
+                message : "Unauthorized"
+            })
+        }
+        
     }catch(err){
         console.log(err);
         return ;
